requirements.txt
openai==0.27.2
typer[all]==0.7.0
pyyaml==6.0.0
tiktoken==0.3.2